feat: Refactor influx classes for stardust/nova (they now share functionaliy in super class InfluxClient). Add daily blocks statistics for nova (api).
