+++ api/src/initServices.ts
@@ -24,12 +24,13 @@ import { LocalStorageService } from "./services/localStorageService";
 import { NetworkService } from "./services/networkService";
 import { ChronicleService as ChronicleServiceNova } from "./services/nova/chronicleService";
 import { NovaFeed } from "./services/nova/feed/novaFeed";
+import { InfluxServiceNova } from "./services/nova/influx/influxServiceNova";
 import { NodeInfoService as NodeInfoServiceNova } from "./services/nova/nodeInfoService";
 import { NovaApiService } from "./services/nova/novaApiService";
 import { NovaStatsService } from "./services/nova/stats/novaStatsService";
 import { ChronicleService as ChronicleServiceStardust } from "./services/stardust/chronicleService";
 import { StardustFeed } from "./services/stardust/feed/stardustFeed";
-import { InfluxDBService } from "./services/stardust/influx/influxDbService";
+import { InfluxServiceStardust } from "./services/stardust/influx/influxServiceStardust";
 import { NodeInfoService as NodeInfoServiceStardust } from "./services/stardust/nodeInfoService";
 import { StardustApiService } from "./services/stardust/stardustApiService";
 import { StardustStatsService } from "./services/stardust/stats/stardustStatsService";
@@ -192,7 +193,7 @@ function initStardustServices(networkConfig: INetwork): void {
     const stardustStatsService = new StardustStatsService(networkConfig);
     ServiceFactory.register(`stats-${networkConfig.network}`, () => stardustStatsService);
 
-    const influxDBService = new InfluxDBService(networkConfig);
+    const influxDBService = new InfluxServiceStardust(networkConfig);
     influxDBService
         .buildClient()
         .then((hasClient) => {
@@ -244,6 +245,17 @@ function initNovaServices(networkConfig: INetwork): void {
             ServiceFactory.register(`feed-${networkConfig.network}`, () => novaFeed);
         });
     });
+
+    const influxDBService = new InfluxServiceNova(networkConfig);
+    influxDBService
+        .buildClient()
+        .then((hasClient) => {
+            logger.debug(`[InfluxDb] Registering client with name "${networkConfig.network}". Has client: ${hasClient}`);
+            if (hasClient) {
+                ServiceFactory.register(`influxdb-${networkConfig.network}`, () => influxDBService);
+            }
+        })
+        .catch((e) => logger.warn(`Failed to build influxDb client for "${networkConfig.network}". Cause: ${e}`));
 }
 
 /**
