+++ api/src/models/influx/nova/IInfluxDbCache.ts
@@ -1,0 +1,17 @@
+import { IBlocksDailyInflux } from "./IInfluxTimedEntries";
+import { DayKey } from "../types";
+
+/**
+ * The cache for influx graphs (daily).
+ */
+export interface IInfluxDailyCache {
+    blocksDaily: Map<DayKey, IBlocksDailyInflux>;
+}
+
+/**
+ * The helper to initialize empty maps
+ * @returns The initial cache object
+ */
+export const initializeEmptyDailyCache = () => ({
+    blocksDaily: new Map(),
+});
