+++ api/src/routes.ts
@@ -294,4 +294,11 @@ export const routes: IRoute[] = [
     },
     { path: "/nova/slot/:network/:slotIndex", method: "get", folder: "nova/slot", func: "get" },
     { path: "/nova/epoch/committee/:network/:epochIndex", method: "get", folder: "nova/epoch/committee", func: "get" },
+    {
+        path: "/nova/analytics/daily/:network",
+        method: "get",
+        folder: "nova/analytics/influx/daily",
+        func: "get",
+        sign: true,
+    },
 ];
