+++ api/src/routes/stardust/analytics/influx/daily/get.ts
@@ -18,9 +18,9 @@ import {
     IUnclaimedGenesisOutputsDailyInflux,
     IUnclaimedTokensDailyInflux,
     IUnlockConditionsPerTypeDailyInflux,
-} from "../../../../../models/influx/IInfluxTimedEntries";
+} from "../../../../../models/influx/stardust/IInfluxTimedEntries";
 import { NetworkService } from "../../../../../services/networkService";
-import { InfluxDBService } from "../../../../../services/stardust/influx/influxDbService";
+import { InfluxServiceStardust } from "../../../../../services/stardust/influx/influxServiceStardust";
 import { ValidationHelper } from "../../../../../utils/validationHelper";
 
 /**
@@ -61,7 +61,7 @@ export async function get(_: IConfiguration, request: INetworkBoundGetRequest): 
         return {};
     }
 
-    const influxService = ServiceFactory.get<InfluxDBService>(`influxdb-${request.network}`);
+    const influxService = ServiceFactory.get<InfluxServiceStardust>(`influxdb-${request.network}`);
 
     return influxService
         ? {
