+++ api/src/services/nova/influx/influxServiceNova.ts
@@ -1,0 +1,50 @@
+import { InfluxDB } from "influx";
+import { BLOCK_DAILY_QUERY } from "./influxQueries";
+import logger from "../../../logger";
+import { INetwork } from "../../../models/db/INetwork";
+import { IInfluxDailyCache, initializeEmptyDailyCache } from "../../../models/influx/nova/IInfluxDbCache";
+import { IBlocksDailyInflux } from "../../../models/influx/nova/IInfluxTimedEntries";
+import { InfluxDbClient } from "../../influx/influxClient";
+
+export class InfluxServiceNova extends InfluxDbClient {
+    /**
+     * The InfluxDb Client.
+     */
+    protected _client: InfluxDB;
+
+    /**
+     * The current influx graphs cache instance.
+     */
+    protected readonly _dailyCache: IInfluxDailyCache;
+
+    /**
+     * The network in context for this client.
+     */
+    protected readonly _network: INetwork;
+
+    constructor(network: INetwork) {
+        super(network);
+        this._dailyCache = initializeEmptyDailyCache();
+    }
+
+    public get blocksDaily() {
+        return this.mapToSortedValuesArray(this._dailyCache.blocksDaily);
+    }
+
+    protected setupDataCollection() {
+        const network = this._network.network;
+        logger.verbose(`[InfluxNova] Setting up data collection for (${network}).`);
+
+        // eslint-disable-next-line no-void
+        void this.collectGraphsDaily();
+    }
+
+    /**
+     * Performs the InfluxDb daily graph data collection.
+     * Populates the dailyCache.
+     */
+    private async collectGraphsDaily() {
+        logger.verbose(`[InfluxNova] Collecting daily stats for "${this._network.network}"`);
+        this.updateCacheEntry<IBlocksDailyInflux>(BLOCK_DAILY_QUERY, this._dailyCache.blocksDaily, "Blocks Daily", true);
+    }
+}
