+++ api/src/services/stardust/influx/influxServiceStardust.ts
@@ -1,5 +1,4 @@
-import { INanoDate, InfluxDB, IPingStats, IResults, toNanoDate } from "influx";
-import moment from "moment";
+import { InfluxDB } from "influx";
 import cron from "node-cron";
 import {
     ADDRESSES_WITH_BALANCE_DAILY_QUERY,
@@ -26,16 +25,15 @@ import {
     MILESTONE_STATS_QUERY_BY_INDEX,
 } from "./influxQueries";
 import logger from "../../../logger";
+import { IMilestoneAnalyticStats } from "../../../models/api/stats/IMilestoneAnalyticStats";
 import { INetwork } from "../../../models/db/INetwork";
 import { SHIMMER } from "../../../models/db/networkType";
 import {
-    DayKey,
-    DAY_KEY_FORMAT,
     IInfluxAnalyticsCache,
     IInfluxDailyCache,
     IInfluxMilestoneAnalyticsCache,
     initializeEmptyDailyCache,
-} from "../../../models/influx/IInfluxDbCache";
+} from "../../../models/influx/stardust/IInfluxDbCache";
 import {
     IAddressesWithBalanceDailyInflux,
     IAliasActivityDailyInflux,
@@ -45,7 +43,6 @@ import {
     INftActivityDailyInflux,
     IOutputsDailyInflux,
     IStorageDepositDailyInflux,
-    ITimedEntry,
     ITokensHeldPerOutputDailyInflux,
     ITokensHeldWithUnlockConditionDailyInflux,
     ITokensTransferredDailyInflux,
@@ -53,7 +50,9 @@ import {
     IUnclaimedGenesisOutputsDailyInflux,
     IUnclaimedTokensDailyInflux,
     IUnlockConditionsPerTypeDailyInflux,
-} from "../../../models/influx/IInfluxTimedEntries";
+} from "../../../models/influx/stardust/IInfluxTimedEntries";
+import { ITimedEntry } from "../../../models/influx/types";
+import { InfluxDbClient } from "../../influx/influxClient";
 
 type MilestoneUpdate = ITimedEntry & {
     milestoneIndex: number;
@@ -83,11 +82,6 @@ const COLLECT_ANALYTICS_DATA_CRON = "55 58 * * * *";
 const COLLECT_SHIMMER_DATA_CRON = "*/30 * * * * *";
 
 /**
- * N of nanoseconds in a millsecond.
- */
-const NANOSECONDS_IN_MILLISECOND = 1000000;
-
-/**
  * Milestone analyitics cache MAX size.
  */
 const MILESTONE_CACHE_MAX = 20;
@@ -95,7 +89,7 @@ const MILESTONE_CACHE_MAX = 20;
 /**
  * The InfluxDb Client wrapper.
  */
-export abstract class InfluxDbClient {
+export class InfluxServiceStardust extends InfluxDbClient {
     /**
      * The InfluxDb Client.
      */
@@ -117,104 +111,130 @@ export abstract class InfluxDbClient {
     protected readonly _milestoneCache: IInfluxMilestoneAnalyticsCache;
 
     /**
-     * The network in context for this client.
-     */
-    private readonly _network: INetwork;
-
-    /**
      * Create a new instance of InfluxDbClient.
      * @param network The network configuration.
      */
     constructor(network: INetwork) {
-        this._network = network;
+        super(network);
         this._dailyCache = initializeEmptyDailyCache();
         this._analyticsCache = {};
         this._milestoneCache = new Map();
     }
 
-    /**
-     * Build a new client instance asynchronously.
-     * @returns Boolean representing that the client ping succeeded.
-     */
-    public async buildClient(): Promise<boolean> {
-        const protocol = this._network.analyticsInfluxDbProtocol || "https";
-        const network = this._network.network;
-        const [host, portString] = this._network.analyticsInfluxDbEndpoint.split(":");
-        // Parse port string to int, or use default port for protocol
-        const port = Number.parseInt(portString, 10) || (protocol === "https" ? 443 : 80);
-        const database = this._network.analyticsInfluxDbDatabase;
-        const username = this._network.analyticsInfluxDbUsername;
-        const password = this._network.analyticsInfluxDbPassword;
-
-        if (host && database && username && password) {
-            logger.verbose(`[InfluxDb] Found configuration for (${network})]`);
-            const token = Buffer.from(`${username}:${password}`, "utf8").toString("base64");
-            const options = {
-                headers: {
-                    Authorization: `Basic ${token}`,
-                },
-            };
-
-            const influxDbClient = new InfluxDB({ protocol, port, host, database, username, password, options });
-
-            return influxDbClient
-                .ping(1500)
-                .then((pingResults: IPingStats[]) => {
-                    if (pingResults.length > 0) {
-                        const anyHostIsOnline = pingResults.some((ping) => ping.online);
-
-                        if (anyHostIsOnline) {
-                            logger.info(`[InfluxDb] Client started for "${network}"...`);
-                            this._client = influxDbClient;
-                            this.setupDataCollection();
-                        }
-
-                        return anyHostIsOnline;
-                    }
-
-                    return false;
-                })
-                .catch((e) => {
-                    logger.verbose(`[InfluxDb] Ping failed for "${network}". ${e}`);
-                    return false;
-                });
-        }
-
-        logger.warn(`[InfluxDb] Configuration not found for "${network}".`);
-        return false;
+    public get blocksDaily() {
+        return this.mapToSortedValuesArray(this._dailyCache.blocksDaily);
     }
 
-    /**
-     * Get the milestone analytics by index and set it in the cache.
-     * @param milestoneIndex - The milestone index.
-     */
-    public async collectMilestoneStatsByIndex(milestoneIndex: number) {
-        try {
-            for (const update of await this._client.query<MilestoneUpdate>(MILESTONE_STATS_QUERY_BY_INDEX, {
-                placeholders: { milestoneIndex },
-            })) {
-                this.updateMilestoneCache(update);
-            }
-        } catch (err) {
-            logger.warn(`[InfluxDb] Failed refreshing milestone stats for "${this._network.network}". Cause: ${err}`);
-        }
+    public get transactionsDaily() {
+        return this.mapToSortedValuesArray(this._dailyCache.transactionsDaily);
     }
 
-    /**
-     * Function to sort map entries in ascending order.
-     * @param a The first entry
-     * @param z The second entry
-     * @returns Negative number if first entry is before second, positive otherwise.
-     */
-    protected readonly ENTRIES_ASC_SORT = (a: ITimedEntry, z: ITimedEntry) => (moment(a.time).isBefore(moment(z.time)) ? -1 : 1);
+    public get outputsDaily() {
+        return this.mapToSortedValuesArray(this._dailyCache.outputsDaily);
+    }
+
+    public get tokensHeldDaily() {
+        return this.mapToSortedValuesArray(this._dailyCache.tokensHeldDaily);
+    }
+
+    public get addressesWithBalanceDaily() {
+        return this.mapToSortedValuesArray(this._dailyCache.addressesWithBalanceDaily);
+    }
+
+    public get activeAddressesDaily() {
+        return this.mapToSortedValuesArray(this._dailyCache.activeAddressesDaily);
+    }
+
+    public get tokensTransferredDaily() {
+        return this.mapToSortedValuesArray(this._dailyCache.tokensTransferredDaily);
+    }
+
+    public get aliasActivityDaily() {
+        return this.mapToSortedValuesArray(this._dailyCache.aliasActivityDaily);
+    }
+
+    public get unlockConditionsPerTypeDaily() {
+        return this.mapToSortedValuesArray(this._dailyCache.unlockConditionsPerTypeDaily);
+    }
+
+    public get nftActivityDaily() {
+        return this.mapToSortedValuesArray(this._dailyCache.nftActivityDaily);
+    }
+
+    public get tokensHeldWithUnlockConditionDaily() {
+        return this.mapToSortedValuesArray(this._dailyCache.tokensHeldWithUnlockConditionDaily);
+    }
+
+    public get unclaimedTokensDaily() {
+        return this.mapToSortedValuesArray(this._dailyCache.unclaimedTokensDaily);
+    }
+
+    public get unclaimedGenesisOutputsDaily() {
+        return this.mapToSortedValuesArray(this._dailyCache.unclaimedGenesisOutputsDaily);
+    }
+
+    public get ledgerSizeDaily() {
+        return this.mapToSortedValuesArray(this._dailyCache.ledgerSizeDaily);
+    }
+
+    public get storageDepositDaily() {
+        return this.mapToSortedValuesArray(this._dailyCache.storageDepositDaily);
+    }
+
+    public get addressesWithBalance() {
+        return this._analyticsCache.addressesWithBalance;
+    }
+
+    public get nativeTokensCount() {
+        return this._analyticsCache.nativeTokensCount;
+    }
+
+    public get nftsCount() {
+        return this._analyticsCache.nftsCount;
+    }
+
+    public get lockedStorageDeposit() {
+        return this._analyticsCache.lockedStorageDeposit;
+    }
+
+    public get totalUnclaimedShimmer() {
+        return this._analyticsCache.totalUnclaimedShimmer;
+    }
+
+    public get milestoneAnalytics() {
+        return this._milestoneCache;
+    }
+
+    public async fetchAnalyticsForMilestone(milestoneIndex: number) {
+        await this.collectMilestoneStatsByIndex(milestoneIndex);
+        return this._milestoneCache.get(milestoneIndex);
+    }
+
+    public async fetchAnalyticsForMilestoneWithRetries(milestoneIndex: number): Promise<IMilestoneAnalyticStats | undefined> {
+        const MAX_RETRY = 30;
+        const RETRY_TIMEOUT = 350;
+
+        let retries = 0;
+        let maybeMsStats = this._milestoneCache.get(milestoneIndex);
+
+        while (!maybeMsStats && retries < MAX_RETRY) {
+            retries += 1;
+            logger.debug(`[InfluxStardust] Try ${retries} of fetching milestone stats for ${milestoneIndex}`);
+            maybeMsStats = this._milestoneCache.get(milestoneIndex);
+
+            await new Promise((f) => setTimeout(f, RETRY_TIMEOUT));
+        }
+
+        return maybeMsStats;
+    }
 
     /**
      * Setup a InfluxDb data collection periodic job.
      * Runs once at the start, then every COLLECT_DATA_FREQ_MS interval.
      */
-    private setupDataCollection() {
+    protected setupDataCollection() {
         const network = this._network.network;
-        logger.verbose(`[InfluxDb] Setting up data collection for (${network})]`);
+        logger.verbose(`[InfluxStardust] Setting up data collection for (${network}).`);
 
         // eslint-disable-next-line no-void
         void this.collectGraphsDaily();
@@ -250,7 +270,7 @@ export abstract class InfluxDbClient {
                 });
             }
         } else {
-            logger.debug(`[InfluxDb] Client not configured for this network "${network}"`);
+            logger.debug(`[InfluxStardust] Client not configured for this network "${network}"`);
         }
     }
 
@@ -259,7 +279,7 @@ export abstract class InfluxDbClient {
      * Populates the dailyCache.
      */
     private async collectGraphsDaily() {
-        logger.verbose(`[InfluxDb] Collecting daily stats for "${this._network.network}"`);
+        logger.verbose(`[InfluxStardust] Collecting daily stats for "${this._network.network}"`);
         this.updateCacheEntry<IBlocksDailyInflux>(BLOCK_DAILY_QUERY, this._dailyCache.blocksDaily, "Blocks Daily");
         this.updateCacheEntry<ITransactionsDailyInflux>(TRANSACTION_DAILY_QUERY, this._dailyCache.transactionsDaily, "Transactions Daily");
         this.updateCacheEntry<IOutputsDailyInflux>(OUTPUTS_DAILY_QUERY, this._dailyCache.outputsDaily, "Outpus Daily");
@@ -322,7 +342,7 @@ export abstract class InfluxDbClient {
      * Populates the analyticsCache.
      */
     private async collectAnalytics() {
-        logger.verbose(`[InfluxDb] Collecting analytic stats for "${this._network.network}"`);
+        logger.verbose(`[InfluxStardust] Collecting analytic stats for "${this._network.network}"`);
         try {
             for (const update of await this.queryInflux<ITimedEntry & { addressesWithBalance: string }>(
                 ADDRESSES_WITH_BALANCE_TOTAL_QUERY,
@@ -356,12 +376,28 @@ export abstract class InfluxDbClient {
                 this._analyticsCache.lockedStorageDeposit = update.lockedStorageDeposit;
             }
         } catch (err) {
-            logger.warn(`[InfluxDb] Failed refreshing analytics for "${this._network.network}"! Cause: ${err}`);
+            logger.warn(`[InfluxStardust] Failed refreshing analytics for "${this._network.network}"! Cause: ${err}`);
+        }
+    }
+
+    /**
+     * Get the milestone analytics by index and set it in the cache.
+     * @param milestoneIndex - The milestone index.
+     */
+    private async collectMilestoneStatsByIndex(milestoneIndex: number) {
+        try {
+            for (const update of await this._client.query<MilestoneUpdate>(MILESTONE_STATS_QUERY_BY_INDEX, {
+                placeholders: { milestoneIndex },
+            })) {
+                this.updateMilestoneCache(update);
+            }
+        } catch (err) {
+            logger.warn(`[InfluxStardust] Failed refreshing milestone stats for "${this._network.network}". Cause: ${err}`);
         }
     }
 
     private async collectShimmerUnclaimed() {
-        logger.verbose(`[InfluxDb] Collecting shimmer stats for "${this._network.network}"`);
+        logger.verbose(`[InfluxStardust] Collecting shimmer stats for "${this._network.network}"`);
 
         try {
             for (const update of await this.queryInflux<ITimedEntry & { totalUnclaimedShimmer: string }>(
@@ -372,18 +408,18 @@ export abstract class InfluxDbClient {
                 this._analyticsCache.totalUnclaimedShimmer = update.totalUnclaimedShimmer;
             }
         } catch (err) {
-            logger.warn(`[InfluxDb] Failed refreshing shimmer stats for "${this._network.network}"! Cause: ${err}`);
+            logger.warn(`[InfluxStardust] Failed refreshing shimmer stats for "${this._network.network}"! Cause: ${err}`);
         }
     }
 
     private async collectMilestoneStats() {
-        logger.debug(`[InfluxDb] Collecting milestone stats for "${this._network.network}"`);
+        logger.debug(`[InfluxStardust] Collecting milestone stats for "${this._network.network}"`);
         try {
             for (const update of await this.queryInflux<MilestoneUpdate>(MILESTONE_STATS_QUERY, null, this.getToNanoDate())) {
                 this.updateMilestoneCache(update);
             }
         } catch (err) {
-            logger.warn(`[InfluxDb] Failed refreshing milestone stats for "${this._network.network}". Cause: ${err}`);
+            logger.warn(`[InfluxStardust] Failed refreshing milestone stats for "${this._network.network}". Cause: ${err}`);
         }
     }
 
@@ -403,7 +439,7 @@ export abstract class InfluxDbClient {
                 },
             });
 
-            logger.debug(`[InfluxDb] Added milestone index "${milestoneIndex}" to cache for "${this._network.network}"`);
+            logger.debug(`[InfluxStardust] Added milestone index "${milestoneIndex}" to cache for "${this._network.network}"`);
 
             if (this._milestoneCache.size > MILESTONE_CACHE_MAX) {
                 let lowestIndex: number;
@@ -417,147 +453,10 @@ export abstract class InfluxDbClient {
                     }
                 }
 
-                logger.debug(`[InfluxDb] Deleting milestone index "${lowestIndex}" ("${this._network.network}")`);
+                logger.debug(`[InfluxStardust] Deleting milestone index "${lowestIndex}" ("${this._network.network}")`);
 
                 this._milestoneCache.delete(lowestIndex);
             }
         }
-    }
-
-    /**
-     * Update one cache entry with InfluxDb data.
-     * Uses the date from the latest entry as FROM timestamp for the update.
-     * @param queryTemplate The query template object.
-     * @param queryTemplate.full Full query (no timespan) and partial (from, to).
-     * @param queryTemplate.partial Parameterized query (from, to).
-     * @param cacheEntryToFetch The cache entry to fetch.
-     * @param description The optional entry description for logging.
-     * @param debug The optional debug boolean to show more logs.
-     */
-    private updateCacheEntry<T extends ITimedEntry>(
-        queryTemplate: { full: string; partial: string },
-        cacheEntryToFetch: Map<DayKey, T>,
-        description: string = "Daily entry",
-        debug: boolean = false,
-    ) {
-        const network = this._network.network;
-        const fromNanoDate: INanoDate | null = this.getFromNanoDate(cacheEntryToFetch);
-
-        if (debug) {
-            logger.debug(
-                `[InfluxDb] Refreshing ${description} from date
-                ${fromNanoDate ? fromNanoDate.toISOString() : null} (${this._network.network})`,
-            );
-        }
-
-        const query = fromNanoDate ? queryTemplate.partial : queryTemplate.full;
-
-        this.queryInflux<T>(query, fromNanoDate, this.getToNanoDate())
-            .then((results) => {
-                for (const update of results) {
-                    if (this.isAnyFieldNotNull<T>(update)) {
-                        if (debug) {
-                            logger.debug(
-                                `[InfluxDb] Setting ${description} cache entry (${network}):`,
-                                moment(update.time).format(DAY_KEY_FORMAT),
-                            );
-                        }
-
-                        cacheEntryToFetch.set(moment(update.time).format(DAY_KEY_FORMAT), update);
-                    } else if (debug) {
-                        logger.warn(
-                            `[InfluxDb] Found empty result entry while populating cache (${network}).
-                            ${JSON.stringify(update)}`,
-                        );
-                    }
-                }
-            })
-            .catch((e) => {
-                logger.warn(`[InfluxDb]] Query ${description} failed for (${network}). Cause ${e}`);
-            });
-    }
-
-    /**
-     * Execute InfluxQL query.
-     * @param query The query.
-     * @param from The starting Date to use in the query.
-     * @param to The ending Date to use in the query.
-     */
-    private async queryInflux<T>(query: string, from: INanoDate | null, to: INanoDate): Promise<IResults<T>> {
-        const params = { placeholders: { from: undefined, to: to.toNanoISOString() } };
-
-        if (from) {
-            params.placeholders.from = from.toNanoISOString();
-        }
-
-        return this._client.query<T>(query, params);
-    }
-
-    /**
-     * Compute the FROM Date for data update query from current cache entry.
-     * @param cacheEntry The current cache entry map.
-     * @returns The (from) (INano)Date.
-     */
-    private getFromNanoDate(cacheEntry: Map<DayKey, ITimedEntry>): INanoDate | null {
-        let fromNanoDate: INanoDate | null;
-        if (cacheEntry.size === 0) {
-            fromNanoDate = null;
-        } else {
-            const lastDate = this.computeLastDateOfContinousSeries(cacheEntry);
-
-            fromNanoDate = toNanoDate(
-                // eslint-disable-next-line newline-per-chained-call
-                (lastDate.hours(0).minutes(0).seconds(1).valueOf() * NANOSECONDS_IN_MILLISECOND).toString(),
-            );
-        }
-
-        return fromNanoDate;
-    }
-
-    /**
-     * Get the last date (day) from cache entry,
-     * which is part of a continous series of days (no holes in the data).
-     * We then use this date and the start for next refresh query, so that if there are holes in the data,
-     * that secion will try to be updated.
-     * @param cacheEntry The current cache entry map.
-     * @returns Moment object representing the latest date of continous data.
-     */
-    private computeLastDateOfContinousSeries(cacheEntry: Map<DayKey, ITimedEntry>): moment.Moment {
-        const sortedEntries = Array.from(cacheEntry.values()).sort(this.ENTRIES_ASC_SORT);
-
-        const oldestEntry = sortedEntries[0];
-        const start = moment(oldestEntry.time);
-        let lastDay = start;
-
-        for (let day = start; day.isBefore(moment(), "day"); day = day.add(1, "day")) {
-            const key = day.format(DAY_KEY_FORMAT);
-
-            if (cacheEntry.has(key)) {
-                lastDay = day;
-            } else {
-                break;
-            }
-        }
-
-        return lastDay;
-    }
-
-    /**
-     * Compute the TO Date for data update query.
-     * @returns Current datetime as INanoDate.
-     */
-    private getToNanoDate(): INanoDate {
-        return toNanoDate((moment().startOf("day").valueOf() * NANOSECONDS_IN_MILLISECOND).toString());
-    }
-
-    /**
-     * Helper function to check if any of the fields (except time) are non-null.
-     * @param data Any object.
-     * @returns True if any of the object fields (excludes time) is not null.
-     */
-    private isAnyFieldNotNull<T>(data: T): boolean {
-        return Object.getOwnPropertyNames(data)
-            .filter((fName) => fName !== "time")
-            .some((fName) => data[fName] !== null);
     }
 }
